beetle/submissions/requirements.txt
torch
torchvision
open-clip-torch